# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: Experiment Management Service
  description: This is the API description of the Experiment Management Service of the Organicity project
  termsOfService: none
  contact:
    name: Organicity
    url: http://organicity.eu/
  license:
    name: none
    url: http://none
host: localhost:8081
...
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:  
...
  /emscheck/experimentrunning/{expId} :
    get:
      tags: 
       - Checks
      description: | 
        Check whether an experiment is running
      operationId: emsCheckExpRunning
      parameters:
        - name: expId
          in: path
          description: Experiment Id
          required: true
          type: string
        - name: Authorization
          in: header
          description: Authorization token
          required: true
          type: string
          format: string
          default: Bearer ...          
      responses:
       "200":
        description: The experiment is running
       "404":
        description: The experiment is not running
       "400":
        description: Bad request
  /emscheck/assets-public/{expId} :
    get:
      tags: 
       - Checks
      description: | 
        Check whether an experiment has public assets
      operationId: emsCheckAssetsPublic
      parameters:
        - name: expId
          in: path
          description: Experiment Id
          required: true
          type: string
        - name: Authorization
          in: header
          description: Authorization token
          required: true
          type: string
          format: string
          default: Bearer ...          
      responses:
       "200":
        description: The experiment has public assets
       "404":
        description: The experiment has not public assets
       "400":
        description: Bad request      
  /emscheck/experimentowner/{experId}/{expId} :
    get:
      tags: 
       - Checks
      description: | 
        Check whether an experimenter is the owner of one experiment
      operationId: emsCheckOwner  
      parameters:
        - name: experId
          in: path
          description: Experimenter Id
          required: true
          type: string
        - name: expId
          type: string
          in: path
          required: true
          description: Experiment ID
        - name: Authorization
          in: header
          description: Authorization token
          required: true
          type: string
          format: string
          default: Bearer ...          
      responses:
       "200":
        description: The experimenter is the owner
       "404":
        description: The experimenter is not the owner
       "400":
        description: Bad request
  /emscheck/application-experiment/{expId}/{appId} :
    get:
      tags: 
       - Checks
      description: | 
        Check whether an application belongs to one experiment
      operationId: emsCheckAppExp
      parameters:
        - name: expId
          in: path
          description: Experiment Id
          required: true
          type: string
        - name: appId
          type: string
          in: path
          required: true
          description: Application ID
        - name: Authorization
          in: header
          description: Authorization token
          required: true
          type: string
          format: string
          default: Bearer ...          
      responses:
       "200":
        description: The application belongs to the experiment
       "404":
        description: The application does not belong to the experiment
       "400":
        description: Bad request      
  /emscheck/participant-experiment/{parId}/{expId} :
    get:
      tags: 
       - Checks
      description: | 
        Check whether a participant takes part in the experiment
      operationId: emsCheckParExp
      parameters:
        - name: expId
          in: path
          description: Experiment Id
          required: true
          type: string
        - name: parId
          type: string
          in: path
          required: true
          description: participant ID
        - name: Authorization
          in: header
          description: Authorization token
          required: true
          type: string
          format: string
          default: Bearer ...          
      responses:
       "200":
        description: The participant takes part in the experiment
       "404":
        description: The participant does not take part in the experiment
       "400":
        description: Bad request        
...
  /experiments/{expid}/datasources:
   get:
      tags: 
       - Data Sources
      description: |
        Get the urns of the assets selected for an experiment
      operationId: dsGet
      parameters:
        - name: Authorization
          in: header
          description: Authorization token
          required: true
          type: string
          format: string
          default: Bearer ...
        - name: expid
          type: string
          in: path
          required: true
          description: Experiment ID
      responses:
        "200":
          description: OK
          schema: 
            $ref: '#/definitions/DataSource'
        "403": 
          description: Authorization failure
definitions:
...
  DataSource:
    required:
      - experimentId
      - experimenterIds
    properties:
      experimentId:
        type: string
      experimenterIds:
        type: array
        items:
          type: string
      urns: 
        type: array
        items:
          type: string
...